crashfix pt2

+++ src/framework/math/box2dtools.cpp
@@ -29,7 +29,7 @@ b2PolygonShape Box2DTools::createBeveledBox(const float width, const float heigh
       b2Vec2(0, height - bevel_offset),
       b2Vec2(0, bevel_offset)};
 
-   box_shape.Set(vertices.data(), vertices.size());
+   box_shape.Set(vertices.data(), static_cast<int32_t>(vertices.size()));
 
    return box_shape;
 }
